Move Values inheritors to Values namespace
